Optimize ExpectationSet.
With the old implementation, looking up the expectations for a single
test was O(n) in the number of status file entries in all status files
(!). That means getting the status of m tests was O(n * m). :(

This parses entry paths into a tree and only traverses the branches of
that the test patch matches. On my script that determines the
expectation of every test on every configuration, it changes the runtime
from ~30s to ~1.5.

On my laptop, this change also takes this:

  time ./tools/test.py -m release --list -c dartdevc language_2

From about ~4.5s to ~3.5.

Storing the expectations in a more structured way also opens the door
to some other changes we could do:

- Separate out the configuration from ExpectationSet so that you can
  use a single ExpectationSet for multiple configurations.

- Determining which status entries are pointless because they are
  covered by a similar expectation higher up the file system tree.

- Determining which status entries collide and hit the same test.

Change-Id: I121bca34b68e1e6bb45edb59bd3cfe95d37e6bef
Reviewed-on: https://dart-review.googlesource.com/38382
Commit-Queue: Bob Nystrom <rnystrom@google.com>
Reviewed-by: William Hesse <whesse@google.com>