Sort assists and fixes by relevance and then by name.
We used to get away with sorting just by relevance because the order
of assists and fixes was correct. I think it is time to switch to
sorting. I still would like to keep 'Surround with XXX' ordered in
a specific way, so I'm changing relevance for these assists.

R=brianwilkerson@google.com, devoncarew@google.com

Change-Id: I77f77ec732d70b31c32164759a3ac62dc790e469
Reviewed-on: https://dart-review.googlesource.com/40880
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>