[kernel] Don't always treat doubles as constants
Previously, when the VM parsed a doubles from the dill file it passed
control over to the constant evaluator which would then case the
value and canonicalize it.
That doesn't really seem to make sense when its not used in a const
context, nor is it done the same way for integers for instance.

This CL changes that and treats doubles the same way as integers.

Below timings are statistics run through "ministat" based on 5 runs
of each configuration (now and before). The timings is for running
the dill file with the VM, i.e. the dill file was precompiled via
fasta.

Program adding 50,000 doubles to a list and prints the last one:

        -56.6 +/- 28.2389
        -10.9988% +/- 5.48753%
  (~510 ms -> ~460 ms)

Program adding 1,000,000 doubles to a list and prints the last one:

        -3177.2 +/- 263.54
        -31.4643% +/- 2.60987%
  (~10 seconds -> ~7 seconds)

Hello world program:

No difference proven at 95.0% confidence

Change-Id: I7dd748e759b3e433bc806b2ef2505d29f0b109b6
Reviewed-on: https://dart-review.googlesource.com/42000
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>