Relax checks in Resolution Applier
If the BodyBuilder assigns a location to an entity, but the analyzer
doesn't, don't regard that as an error.

This can happen, for instance, when the BodyBuilder creates a
synthetic `null` to refer to the default value of an optional
parameter.  For better debuggability, it makes sense for the
BodyBuilder to assign a location for that synthetic `null`, but there
is no corresponding analyzer token.

Change-Id: I91bb77bfe900a834353bf90482932321fecde416
Reviewed-on: https://dart-review.googlesource.com/42200
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>