Convert paths for mock SDK libraries to fix tests on Windows
The other lines on this method are calling `provider.convertPath` around
their paths but these ones do not, which is causing almost all analysis
server tests to fail on Windows (for me - I can't explain them not
failing builds). This causes the hard-coded paths like
`/lib/core/core.dart` to be fixed up to `C:\lib\core\core.dart` which is
what the server is looking for (because at the other end the paths
*have* been fixed up).

See #32226.