Fix refactoring tests on Windows
This hand-built JSON fails on Windows because the paths being inserted
only have a single slash ("C:\test.dart") but in the real results the
JSON is encoded correctly and hhas two paths.

This works around the issue by JSON-encoding the filename but really the
whole object should be serialised to JSON properly.

Bug: https://github.com/dart-lang/sdk/issues/32226
Change-Id: I8aa418e8e3e6377ed01fe3bbecc3242f3a012629
Reviewed-on: https://dart-review.googlesource.com/42742
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>