Remove $runtime != none sections from DDC status files
These sections are no longer necessary (and neither is $runtime==none)
after test.py was changed to run tests with compile-time errors on
browser runtimes and to not expect a runtime error from a test that
has no runtime.

The status OK doesn't work for tests that have (only) a runtime error
expectation, because removing the runtime error and leaving just OK is
not okay.

Change-Id: I97998e0d75dfd1c5e01f2f80368416dcf0303114
Reviewed-on: https://dart-review.googlesource.com/43661
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Jenny Messerly <jmesserly@google.com>