[fasta] Incremental compilation fallback
In the incremental compiler we can bootstrap from a dill file.
In some cases this can lead to a crash, where compiling from scratch
wouldn't have crashed.
This CL introduces a fallback, so that if compilation crashes
(technically when we build the outline) when bootstrapping from a dill
file, we retry without bootstrapping from the dill file.

Fixes 32384.

Change-Id: Iaefc01a2c03589f615d34910b7b93f73f1fbdc36
Reviewed-on: https://dart-review.googlesource.com/44720
Reviewed-by: Peter von der Ahé <ahe@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>