[VM] Guard against duplicate entries in the static [Redirection::list_]
The ARM simulator has assertions that there are no duplicates in the
[Redirection::list_] but the code does not guarantee this, since
multiple threads can add the same add [Redirection] entries for the same
native concurrently.

This CL guards the modification of [list_] with a [Mutex] and ensures
there's no duplicates.

This issue caused flakiness on the debug-simarm configuration.

Change-Id: Ia84a08e6e6f851a6047bae047ffb990044fb4a4b
Reviewed-on: https://dart-review.googlesource.com/45281
Reviewed-by: Erik Corry <erikcorry@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>