Update status after co19 roll
The only change in co19-runtime.status that isn't just combining two
sections and normalizing is the movement of

Language/Classes/definition_t24: MissingCompileTimeError # Issue co19 #129, sdk issue #31624

from the vm section to a section with vm, dart_precompiled, and flutter, and
the removal of two sections only containing a single Generative_Constructors
test each.

The file is also normalized.

Bug: dartbug.com/32434
Change-Id: Ic62817b3ebdfd27e3cd58abc21f8bd61e7a18664
Reviewed-on: https://dart-review.googlesource.com/45380
Reviewed-by: Peter von der Ahé <ahe@google.com>