Store method invocation arguments in summaries when needed for type inference.
When summarizing an initializer expression that contains a method
call, we need to store the method arguments in case they are needed
for type inference.  They might be needed if the method in question is
a generic method and no type arguments are provided at the call site.

Since we do serialization based on an unresolved AST, we can't tell if
the method in question is generic at serialization time, so we
serialize the arguments of any method invocation that doesn't specify
explicit type arguments.

Addresses one of the root causes of #32394

Change-Id: Ibb50bdfc5ed29a518f51b13d2ffd72c484ab461d
Reviewed-on: https://dart-review.googlesource.com/46301
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>