[VM] When generating the token position list for each line (for a [Script]), only scan relevant classes/functions/fields
The source-based pipeline was simply iterating the token script of a
[Script] to generate a list of (line, [token positions]) tuples.

The kernel-based pipeline is not based on the source (and therefore also
not based on the token stream).  Instead it's based on information from
two places:

  a) The .dill file constains a delta-encoded list of line start file
     offsets.  This line starts array is attached to [Script] objects via
     [RawScript.line_starts_].

  b) It scans the *entire* isolate's program structure (i.e. libraries,
     classes, functions and fields) and does abstract interpretation
     of the kernel blob (ast) and saves the encountered token positions
     on the side.  Afterwards the token positions are merged together
     with the line-starts from a) which then produces the list of (line,
     token positions) tuples.

This CL changes b) to only do the abstract interpretation of
functions/fields which are relevant for the [Script] we are generating
the list of (line, [token position]) tuples for.

Nonetheless, b) will still iterate over the entire isolate's program
structure to find all possible elements belonging to a specific
[Script].  This is a separate issue which should be handled.


On one example it reduces time to get the coverage from ~ 35 seconds to 9.5
seconds (baseline from dart-v1 is around 6.5 seconds).

Issue https://github.com/dart-lang/sdk/issues/32562

Change-Id: I64cda28666c732938379cdf4b2ac62a2371cb3b7
Reviewed-on: https://dart-review.googlesource.com/47080
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>